requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
plotly>=5.13.0
numpy>=1.24.0
gitpython>=3.1.30
python-magic>=0.4.27
pathlib>=1.0.1
typing>=3.7.4
python-dotenv>=1.0.0
Pygments>=2.17.2
zipfile38==0.0.3
radon>=6.0.1
pylint>=3.0.2
astroid>=3.0.1
pyarrow>=15.0.0
requests>=2.31.0
openai>=1.13.3
anthropic>=0.18.1
google-generativeai>=0.3.2
cohere>=4.51
llama-cpp-python>=0.2.68 